Add class Console to refactor class App
